- declared and installed a high-level functor PreCompose - (re)installed two 2-argument PreCompose methods - declared PreComposeMorphisms - simplified \* by moving the check AreComposableMorphisms   to _Functor_AsATwoSequence_OnObjects
We realized that <C>PreCompose</C> has to be installed for each Abelian
category depending on the two morphisms rather than on a single
(<C>IsHomalgComplex</C> and <C>IsATwoSequence</C>);
this is necessary for a clean and uncomplicated method selection.
But in doing so, we lost the caching provided by the functor
<C>AsATwoSequence</C> which is used in other places of the &homalg; code,
and the number of external calls jumped due to unnecessary repeated
computations:
For that reason this high-level <C>PreCompose</C> functor was
reintroduced here, which takes an <C>IsHomalgComplex and IsATwoSequence</C>
as its single argument and then calls a low-level
<C>PreComposeMorphisms</C> method. <C>PreComposeMorphisms</C> must be
the name of the operation of the specific <C>PreCompose</C> functor
in a package implementing a specific Abelian category.
Besides, two <C>PreCompose</C> methods depending on two morphisms were
installed below, which could be overloaded by each specific
<C>PreCompose</C> functor if it uses <C>PreCompose</C>
instead of <C>PreComposeMorphisms</C> as the name of its operation
(and hence bypassing this high-level functor
 and the <C>AsATwoSequence</C> functor).
So by choosing between <C>PreComposeMorphisms</C> and <C>PreCompose</C>
as operation name the authors of the specific packages
have the freedom to choose if they would like to make use
of the caching provided by <C>AsATwoSequence</C> or not.
For this to work, only the two-argument methods for <C>PreCompose</C>
may be called.