major changes as listed below:
Rewritten most parts of of HomogeneousExteriorComplexToModules and some changes to GlobalSetionsModule, TateResolution, LinearStrandOfTateResolution and KoszulRightAdjoint.
These changes removed the identifications of objects and added a few natural transformations (which are modelled as attributes).
In this process got rid of many maps constructed via matrices and replaces this by maps constructed via the functors BaseChange and HomogeneousPartOverCoefficientsRing.

New non-trivial example in doc for global section modules.
Added other examples to the folder (some of them due to Mohamed).

Removed NaturalMapToGlobalSectionsModule, which up to now will not work due to removed identifications.

Added some heuristics to GlobalSectionsModule (this became possible, since we do not need identifications).
In particular, graded free modules and modules with regularity <=0 are treated much faster.

The TateResolution and LinearStrandOfTateResolution are stored with respect to the presentation.
This will be necessary until functors can use the merging of results of different presentation.

Updated examples according to above changes.