Added constructors and methods for GradedModules
Some constructors were added, and all constructors were
redesigned to work with module elements as degrees.
You are still able to set degrees as integers, but you
may take care that the lists or integers match to the
DegreeGroup of the ring, and all degrees are internally handled
as module elements.