Reparied a bug in GeneratorsOfHomogeneousPart
Beforehand we had a heuristical approchad that was used when the target degree was the minimal degree in the module (for positively graded rings)
Then we just took that generators.
But this didn't need to be a basis (only in the case when these generators had no relations).
Now we check whether these generators have relations by checking the degrees of the relations and only apply this heuristics when there are no relations.
Drawback: That might cause a call for NonTrivialDegreesOfXXX.