Return list expressions instead of lists in Singular
Using the current method, B and T have to be copied twice: into the list l
inside the procedure and out of homalg_variable_l after the procedure. The
new method avoids this. This reduces peak RAM usage (and also runtime) and
fixes the "memory leak" that homalg_variable_l is kept alive indefinitely.